+++ src/include/ipxe/efi/efi_autoboot.h
@@ -9,6 +9,8 @@
 
 FILE_LICENCE ( GPL2_OR_LATER_OR_UBDL );
 
-extern void efi_set_autoboot ( void );
+#include <ipxe/efi/efi.h>
+
+extern int efi_set_autoboot_ll_addr ( EFI_HANDLE device );
 
 #endif /* _IPXE_EFI_AUTOBOOT_H */
