+++ src/include/ipxe/errfile.h
@@ -385,6 +385,7 @@ FILE_LICENCE ( GPL2_OR_LATER_OR_UBDL );
 #define ERRFILE_ntlm		      ( ERRFILE_OTHER | 0x00510000 )
 #define ERRFILE_efi_veto	      ( ERRFILE_OTHER | 0x00520000 )
 #define ERRFILE_efi_autoboot	      ( ERRFILE_OTHER | 0x00530000 )
+#define ERRFILE_efi_autoexec	      ( ERRFILE_OTHER | 0x00540000 )
 
 /** @} */
